refactor: /rootにあったものを/chatに移しコード間の隙間を少し開けた
