+++ web/src/features/routes/chat/footer.module.scss
