+++ web/src/features/routes/chat/footer.tsx
@@ -1,14 +1,14 @@
 "use client";
 import { Button } from "@/components/ui/button";
 import { Textarea } from "@/components/ui/textarea";
-import style from "@routes/root/footer.module.scss";
+import style from "./footer.module.scss";
 import { useAtom } from "jotai";
 import { SendHorizontal } from "lucide-react";
 import { type RefObject, useRef, useState } from "react";
 import type { KeyboardEvent } from "react";
 import { v4 as uuidv4 } from "uuid";
-import { messagesAtom } from "../../message/store";
-import type { Message } from "../../message/type";
+import { messagesAtom } from "@/features/message/store";
+import type { Message } from "@/features/message/type";
 
 type Props = { bottomRef: RefObject<HTMLDivElement | null> };
 
@@ -21,9 +21,11 @@ export const Footer = (props: Props) => {
     if (!inputRef.current) {
       return;
     }
+
     if (inputRef.current.value === "") {
       return;
     }
+
     const newMessage: Message = {
       id: uuidv4(),
       author: "me",
@@ -31,10 +33,12 @@ export const Footer = (props: Props) => {
       is_me: true,
       icon: "https://github.com/shadcn.png"
     };
+
     inputRef.current.value = "";
     setTimeout(() => {
       props.bottomRef.current?.scrollIntoView();
     }, 0);
+
     const yesMessage: Message = {
       id: uuidv4(),
       author: "bot",
@@ -42,17 +46,21 @@ export const Footer = (props: Props) => {
       is_me: false,
       icon: "https://github.com/shadcn.png"
     };
+
     setMessages([...messages, newMessage, yesMessage]);
   };
+
   const handleKeyDown = (event: KeyboardEvent<HTMLTextAreaElement>) => {
     if (event.shiftKey) {
       return;
     }
+
     if (event.key === "Enter" && !isComposing) {
       event.preventDefault();
       handleSubmit();
     }
   };
+
   return (
     <div className={style.footer}>
       <Textarea
